move SequencerController to the library instead of the example2 project
